Compatibility changes
Created a new event for sending emails to make backwards compat with renewed's qb-phone data input.

Updated snipe-menu.isAdmin() checks to be performed by ps-adminmenu.

Updated the phone to recognise when a player is no longer knocked out or dead to re-enable the phone.

Updated summit_lib checkjobgrade to a qb-core function.

Edited DoesPlayerExist to just check instead of looking for a function in QBCore to check `DoesPlayerExist`.

Updated summit_bank to renewed banking system

Updated transfernumbers migration command to easily find the phone in the players table under charinfo.

Updated garage data to use base images of the vehicles instead of a category image.