Updated Password fields
I've updated the password fields in;
DarkChat
Mail
Pigeon

These now block out the password when the user is typing them in and adds a "eye" UX element that allows the user to toggle between seeing the password field box and blocking it out.

To Come;
Updated UX so pressing "ENTER" on the keyboard will parse the input fields so the user doesn't have to click the button on the phone to login.